cloud_run/bq-functions/requirements.txt
fastapi[standard]
pydantic
uvicorn[standard]
gunicorn
newspaper4k
lxml[html_clean]
typing_extensions